# Stockwave-Postman-Collection-Json2OpenApi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stockwave",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "variables": {}
    }
  ],
  "paths": {
    "/aluno": {
      "get": {
        "tags": [
          "Aluno"
        ],
        "summary": "Listar Alunos",
        "operationId": "ListarAlunos",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Aluno"
        ],
        "summary": "Cadastrar Aluno",
        "operationId": "CadastrarAluno",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CadastrarAlunoRequest"
                  },
                  {
                    "example": {
                      "cpf_usuario": "25328392015",
                      "email_usuario": "EMAIL@EMAILa.COM.BR",
                      "nome_usuario": "ALUNO TESTE 2",
                      "dt_nasc_aluno": "2000-07-12",
                      "dt_reg_aluno": "2023-04-01",
                      "moedas_aluno": 0,
                      "nivel_aluno": {
                        "nome_nivel": "EASY"
                      },
                      "senha_aluno": "senha2"
                    }
                  }
                ]
              },
              "example": {
                "cpf_usuario": "25328392015",
                "email_usuario": "EMAIL@EMAILa.COM.BR",
                "nome_usuario": "ALUNO TESTE 2",
                "dt_nasc_aluno": "2000-07-12",
                "dt_reg_aluno": "2023-04-01",
                "moedas_aluno": 0,
                "nivel_aluno": {
                  "nome_nivel": "EASY"
                },
                "senha_aluno": "senha2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/aluno/2": {
      "get": {
        "tags": [
          "Aluno"
        ],
        "summary": "Exibir Aluno",
        "operationId": "ExibirAluno",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "Aluno"
        ],
        "summary": "Atualizar Aluno",
        "operationId": "AtualizarAluno",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AtualizarAlunoRequest"
                  },
                  {
                    "example": {
                      "cpf_usuario": "25328392015",
                      "email_usuario": "EMAIL@EMAasasas.COM.BR",
                      "id_usuario": 2,
                      "nome_usuario": "ALUNO TESTE 2",
                      "dt_nasc_aluno": "2000-07-14",
                      "dt_reg_aluno": "2023-04-01",
                      "moedas_aluno": 0,
                      "nivel_aluno": {
                        "nome_nivel": "EASY"
                      },
                      "senha_aluno": "senha2"
                    }
                  }
                ]
              },
              "example": {
                "cpf_usuario": "25328392015",
                "email_usuario": "EMAIL@EMAasasas.COM.BR",
                "id_usuario": 2,
                "nome_usuario": "ALUNO TESTE 2",
                "dt_nasc_aluno": "2000-07-14",
                "dt_reg_aluno": "2023-04-01",
                "moedas_aluno": 0,
                "nivel_aluno": {
                  "nome_nivel": "EASY"
                },
                "senha_aluno": "senha2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/aluno/login": {
      "post": {
        "tags": [
          "Aluno"
        ],
        "summary": "Validar Login Aluno",
        "operationId": "ValidarLoginAluno",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ValidarLoginAlunoRequest"
                  },
                  {
                    "example": {
                      "email_usuario": "EMAIl@EMAILA.coM.BR1",
                      "senha_aluno": "senha2"
                    }
                  }
                ]
              },
              "example": {
                "email_usuario": "EMAIl@EMAILA.coM.BR1",
                "senha_aluno": "senha2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/aula": {
      "get": {
        "tags": [
          "Aula"
        ],
        "summary": "Listar Aulas",
        "operationId": "ListarAulas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Aula"
        ],
        "summary": "Cadastrar Aula",
        "operationId": "CadastrarAula",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CadastrarAulaRequest"
                  },
                  {
                    "example": {
                      "conteudo_aula": "TESTE",
                      "descricao_aula": "TESTE",
                      "id_aula": 3,
                      "nome_aula": "TESTE",
                      "url_audio_aula": "HTTP://WWW.AUDIO.COM/TESTE",
                      "url_video_aula": "HTTP://WWW.VIDEO.COM/TESTE"
                    }
                  }
                ]
              },
              "example": {
                "conteudo_aula": "TESTE",
                "descricao_aula": "TESTE",
                "id_aula": 3,
                "nome_aula": "TESTE",
                "url_audio_aula": "HTTP://WWW.AUDIO.COM/TESTE",
                "url_video_aula": "HTTP://WWW.VIDEO.COM/TESTE"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/aula/2": {
      "get": {
        "tags": [
          "Aula"
        ],
        "summary": "Exibir Aula",
        "operationId": "ExibirAula",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "Aula"
        ],
        "summary": "Atualizar Aula",
        "operationId": "AtualizarAula",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AtualizarAulaRequest"
                  },
                  {
                    "example": {
                      "conteudo_aula": "O QUE É IPO",
                      "descricao_aula": "AULA EXPLICANDO IPO",
                      "id_aula": 2,
                      "nome_aula": "AULA 02",
                      "url_audio_aula": "HTTP://WWW.AUDIO.COM/AULA02",
                      "url_video_aula": "HTTP://WWW.VIDEO.COM/AULA02"
                    }
                  }
                ]
              },
              "example": {
                "conteudo_aula": "O QUE É IPO",
                "descricao_aula": "AULA EXPLICANDO IPO",
                "id_aula": 2,
                "nome_aula": "AULA 02",
                "url_audio_aula": "HTTP://WWW.AUDIO.COM/AULA02",
                "url_video_aula": "HTTP://WWW.VIDEO.COM/AULA02"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "AtualizarAlunoRequest": {
        "title": "AtualizarAlunoRequest",
        "required": [
          "cpf_usuario",
          "email_usuario",
          "id_usuario",
          "nome_usuario",
          "dt_nasc_aluno",
          "dt_reg_aluno",
          "moedas_aluno",
          "nivel_aluno",
          "senha_aluno"
        ],
        "type": "object",
        "properties": {
          "cpf_usuario": {
            "type": "string"
          },
          "email_usuario": {
            "type": "string"
          },
          "id_usuario": {
            "type": "integer",
            "format": "int32"
          },
          "nome_usuario": {
            "type": "string"
          },
          "dt_nasc_aluno": {
            "type": "string"
          },
          "dt_reg_aluno": {
            "type": "string"
          },
          "moedas_aluno": {
            "type": "integer",
            "format": "int32"
          },
          "nivel_aluno": {
            "$ref": "#/components/schemas/NivelAluno"
          },
          "senha_aluno": {
            "type": "string"
          }
        },
        "example": {
          "cpf_usuario": "25328392015",
          "email_usuario": "EMAIL@EMAasasas.COM.BR",
          "id_usuario": 2,
          "nome_usuario": "ALUNO TESTE 2",
          "dt_nasc_aluno": "2000-07-14",
          "dt_reg_aluno": "2023-04-01",
          "moedas_aluno": 0,
          "nivel_aluno": {
            "nome_nivel": "EASY"
          },
          "senha_aluno": "senha2"
        }
      },
      "NivelAluno": {
        "title": "NivelAluno",
        "required": [
          "nome_nivel"
        ],
        "type": "object",
        "properties": {
          "nome_nivel": {
            "type": "string"
          }
        },
        "example": {
          "nome_nivel": "EASY"
        }
      },
      "CadastrarAlunoRequest": {
        "title": "CadastrarAlunoRequest",
        "required": [
          "cpf_usuario",
          "email_usuario",
          "nome_usuario",
          "dt_nasc_aluno",
          "dt_reg_aluno",
          "moedas_aluno",
          "nivel_aluno",
          "senha_aluno"
        ],
        "type": "object",
        "properties": {
          "cpf_usuario": {
            "type": "string"
          },
          "email_usuario": {
            "type": "string"
          },
          "nome_usuario": {
            "type": "string"
          },
          "dt_nasc_aluno": {
            "type": "string"
          },
          "dt_reg_aluno": {
            "type": "string"
          },
          "moedas_aluno": {
            "type": "integer",
            "format": "int32"
          },
          "nivel_aluno": {
            "$ref": "#/components/schemas/NivelAluno"
          },
          "senha_aluno": {
            "type": "string"
          }
        },
        "example": {
          "cpf_usuario": "25328392015",
          "email_usuario": "EMAIL@EMAILa.COM.BR",
          "nome_usuario": "ALUNO TESTE 2",
          "dt_nasc_aluno": "2000-07-12",
          "dt_reg_aluno": "2023-04-01",
          "moedas_aluno": 0,
          "nivel_aluno": {
            "nome_nivel": "EASY"
          },
          "senha_aluno": "senha2"
        }
      },
      "ValidarLoginAlunoRequest": {
        "title": "ValidarLoginAlunoRequest",
        "required": [
          "email_usuario",
          "senha_aluno"
        ],
        "type": "object",
        "properties": {
          "email_usuario": {
            "type": "string"
          },
          "senha_aluno": {
            "type": "string"
          }
        },
        "example": {
          "email_usuario": "EMAIl@EMAILA.coM.BR1",
          "senha_aluno": "senha2"
        }
      },
...
      "AtualizarAulaRequest": {
        "title": "AtualizarAulaRequest",
        "required": [
          "conteudo_aula",
          "descricao_aula",
          "id_aula",
          "nome_aula",
          "url_audio_aula",
          "url_video_aula"
        ],
        "type": "object",
        "properties": {
          "conteudo_aula": {
            "type": "string"
          },
          "descricao_aula": {
            "type": "string"
          },
          "id_aula": {
            "type": "integer",
            "format": "int32"
          },
          "nome_aula": {
            "type": "string"
          },
          "url_audio_aula": {
            "type": "string"
          },
          "url_video_aula": {
            "type": "string"
          }
        },
        "example": {
          "conteudo_aula": "O QUE É IPO",
          "descricao_aula": "AULA EXPLICANDO IPO",
          "id_aula": 2,
          "nome_aula": "AULA 02",
          "url_audio_aula": "HTTP://WWW.AUDIO.COM/AULA02",
          "url_video_aula": "HTTP://WWW.VIDEO.COM/AULA02"
        }
      },
      "CadastrarAulaRequest": {
        "title": "CadastrarAulaRequest",
        "required": [
          "conteudo_aula",
          "descricao_aula",
          "id_aula",
          "nome_aula",
          "url_audio_aula",
          "url_video_aula"
        ],
        "type": "object",
        "properties": {
          "conteudo_aula": {
            "type": "string"
          },
          "descricao_aula": {
            "type": "string"
          },
          "id_aula": {
            "type": "integer",
            "format": "int32"
          },
          "nome_aula": {
            "type": "string"
          },
          "url_audio_aula": {
            "type": "string"
          },
          "url_video_aula": {
            "type": "string"
          }
        },
        "example": {
          "conteudo_aula": "TESTE",
          "descricao_aula": "TESTE",
          "id_aula": 3,
          "nome_aula": "TESTE",
          "url_audio_aula": "HTTP://WWW.AUDIO.COM/TESTE",
          "url_video_aula": "HTTP://WWW.VIDEO.COM/TESTE"
        }
      },
...
    }
  },
...
}